# server/src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ErrorCodes": {
				"enum": [
					"NOT_ENOUGH_CARDINAL",
					"BAD_ADDRESS_VALUE",
					"BAD_INSCRIPTION_VALUE",
					"INSCRIPTION_NOT_IN_WALLET",
					"NOT_YOUR_WALLET",
					"USER_NAME_UNAVAILABLE",
					"BAD_LOGIN_ATTEMPT",
					"NO_TOKEN",
					"NOT_ALLOWED"
				],
				"type": "string"
			},
			"CardinalBalance": {
				"properties": {
					"cardinal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cardinal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Inscription": {
				"description": "Inscription query result",
				"properties": {
					"inscription": {
						"type": "string",
						"description": "Example: \"bebefd42cf76983bdd24df8dfd565b816a049ef78b1dd5c5797b7be3bb8fd9c1i0\""
					},
					"location": {
						"type": "string",
						"description": "Example: \"bebefd42cf76983bdd24df8dfd565b816a049ef78b1dd5c5797b7be3bb8fd9c1:0:0\""
					},
					"explorer": {
						"type": "string",
						"description": "Example: \"https://ordinals.com/inscription/bebefd42cf76983bdd24df8dfd565b816a049ef78b1dd5c5797b7be3bb8fd9c1i0\""
					}
				},
				"required": [
					"inscription",
					"location",
					"explorer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCTransactions": {
				"properties": {
					"transaction": {
						"type": "string"
					},
					"confirmations": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"transaction",
					"confirmations"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Ord": {
				"properties": {
					"inscription_number": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"inscription_number",
					"id"
				],
				"type": "object",
				"description": "Model Ord"
			},
			"WalletMnemonic": {
				"properties": {
					"mnemonic": {
						"type": "string",
						"description": "Example: \"forum now icon hunt best ensure social muffin slab travel perfect burger\""
					}
				},
				"required": [
					"mnemonic"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCWallet": {
				"properties": {
					"id": {
						"type": "string",
						"description": "Name of the wallet, this will create a folder in the .bitcoin/wallets location."
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCTransaction": {
				"description": "Represents a raw, unsigned bitcoin transactions",
				"properties": {
					"version": {
						"type": "number",
						"format": "double"
					},
					"lock_time": {
						"type": "number",
						"format": "double"
					},
					"input": {
						"items": {
							"properties": {
								"witness": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"sequence": {
									"type": "number",
									"format": "double"
								},
								"script_sig": {
									"type": "string"
								},
								"previous_output": {
									"type": "string"
								}
							},
							"required": [
								"witness",
								"sequence",
								"script_sig",
								"previous_output"
							],
							"type": "object"
						},
						"type": "array"
					},
					"output": {
						"items": {
							"properties": {
								"script_pubkey": {
									"type": "string"
								},
								"value": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"script_pubkey",
								"value"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"version",
					"lock_time",
					"input",
					"output"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InscriptionCreationResult": {
				"description": "Inscription creation result.",
				"properties": {
					"commit_tx": {
						"$ref": "#/components/schemas/BTCTransaction"
					},
					"reveal_tx": {
						"$ref": "#/components/schemas/BTCTransaction"
					},
					"commit": {
						"type": "string",
						"description": "Example: \"43a719dc2a5113df2a59c9b84a28e9c8a6f880e7c88c89a61f0980fe20843cb0\""
					},
					"inscription": {
						"type": "string",
						"description": "Example: \"81171dd249c959afec82221ad7467cf52b37f76a1d78e0ed2999528937b9340bi0\";"
					},
					"reveal": {
						"type": "string",
						"description": "Example: \"81171dd249c959afec82221ad7467cf52b37f76a1d78e0ed2999528937b9340b\";"
					},
					"fees": {
						"type": "number",
						"format": "double",
						"description": "Example: 9360;"
					}
				},
				"required": [
					"commit",
					"inscription",
					"reveal",
					"fees"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BTCAddress": {
				"properties": {
					"address": {
						"type": "string"
					}
				},
				"required": [
					"address"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			}
		}
	},
	"info": {
		"title": "ordinals-wallet-api",
		"version": "0.0.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/authentication/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorCodes"
										},
										{
											"properties": {
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"transactions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/BTCTransactions"
															},
															"type": "array"
														}
													]
												},
												"inscriptions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/Inscription"
															},
															"type": "array"
														}
													]
												},
												"balance": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"$ref": "#/components/schemas/CardinalBalance"
														}
													]
												},
												"token": {}
											},
											"required": [
												"addresses",
												"transactions",
												"inscriptions",
												"balance",
												"token"
											],
											"type": "object"
										},
										{
											"properties": {
												"transactions": {},
												"inscriptions": {},
												"balance": {},
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"token": {}
											},
											"required": [
												"addresses",
												"token"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/authentication/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorCodes"
										},
										{
											"properties": {
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"transactions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/BTCTransactions"
															},
															"type": "array"
														}
													]
												},
												"inscriptions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/Inscription"
															},
															"type": "array"
														}
													]
												},
												"balance": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"$ref": "#/components/schemas/CardinalBalance"
														}
													]
												},
												"token": {}
											},
											"required": [
												"addresses",
												"transactions",
												"inscriptions",
												"balance",
												"token"
											],
											"type": "object"
										},
										{
											"properties": {
												"transactions": {},
												"inscriptions": {},
												"balance": {},
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"token": {}
											},
											"required": [
												"addresses",
												"token"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ordinals": {
			"get": {
				"operationId": "OrdIndex",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/Ord"
											},
											"type": "array"
										},
										{
											"type": "string",
											"enum": [
												"Sorry, your request is too powerful :(. Ask for a maximum of 50 ordinals at a time."
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"ord"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 2,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/wallet/status": {
			"get": {
				"operationId": "Status",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [],
				"parameters": []
			}
		},
		"/wallet/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"btcWallet": {
													"$ref": "#/components/schemas/BTCWallet"
												},
												"walletMnemonic": {
													"$ref": "#/components/schemas/WalletMnemonic"
												}
											},
											"required": [
												"btcWallet",
												"walletMnemonic"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorCodes"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/wallet/restore": {
			"post": {
				"operationId": "Restore",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/wallet/balance": {
			"get": {
				"operationId": "Balance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CardinalBalance"
										},
										{
											"$ref": "#/components/schemas/ErrorCodes"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/wallet/transactions": {
			"get": {
				"operationId": "Transactions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorCodes"
										},
										{
											"items": {
												"$ref": "#/components/schemas/BTCTransactions"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/wallet/inscriptions": {
			"get": {
				"operationId": "Inscriptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorCodes"
										},
										{
											"items": {
												"$ref": "#/components/schemas/Inscription"
											},
											"type": "array"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/wallet/inscribe": {
			"post": {
				"operationId": "Inscribe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/InscriptionCreationResult"
										},
										{
											"$ref": "#/components/schemas/ErrorCodes"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"dryRun": {
										"type": "string"
									},
									"feeRate": {
										"type": "string"
									}
								},
								"required": [
									"file",
									"feeRate"
								]
							}
						}
					}
				}
			}
		},
		"/wallet/receive": {
			"get": {
				"operationId": "Receive",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/BTCAddress"
										},
										{
											"$ref": "#/components/schemas/ErrorCodes"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/wallet/send": {
			"post": {
				"operationId": "Send",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/BTCTransaction"
										},
										{
											"$ref": "#/components/schemas/ErrorCodes"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"feeRate": {
										"type": "number",
										"format": "double"
									},
									"inscription": {
										"type": "string"
									},
									"addressTo": {
										"type": "string"
									}
								},
								"required": [
									"feeRate",
									"inscription",
									"addressTo"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/wallet/data": {
			"post": {
				"operationId": "GetWalletData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"transactions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/BTCTransactions"
															},
															"type": "array"
														}
													]
												},
												"inscriptions": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"items": {
																"$ref": "#/components/schemas/Inscription"
															},
															"type": "array"
														}
													]
												},
												"balance": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/ErrorCodes"
														},
														{
															"$ref": "#/components/schemas/CardinalBalance"
														}
													]
												},
												"token": {}
											},
											"required": [
												"addresses",
												"transactions",
												"inscriptions",
												"balance",
												"token"
											],
											"type": "object"
										},
										{
											"properties": {
												"transactions": {},
												"inscriptions": {},
												"balance": {},
												"addresses": {
													"items": {
														"type": "string"
													},
													"type": "array"
												},
												"token": {}
											},
											"required": [
												"addresses",
												"token"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"wallet"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}